Move ScaleBar math functions to Math.svelte.ts
