+++ src/lib/ScaleBarHelper.svelte.ts
@@ -1,4 +1,6 @@
 import { untrack } from "svelte";
+import { getLargestLessThanOrEqualTo, getScientific } from "./Math.svelte";
+import type { MetricUnit, NumberScientific } from "./Math.svelte";
 
 // max length in pixels (length max pixels)
 //  get real-world max length in meters (length max meters)
@@ -27,48 +29,7 @@ import { untrack } from "svelte";
 // Scale bar in base units (meters). Conversion to pixels is a separate concern.
 // Give a maximum length in base units (meter), get a largest-less-than length in base and display units
 
-interface NumberScientific {
-    significand: number,
-    exponent: number,
-}
-function getScientific(num: number): NumberScientific {
-    let significand;
-    let exponent;
-    [significand, exponent] = num.toExponential().split('e').map(Number);
-    return {significand, exponent};
-}
-interface MetricUnit {
-    symbol: string,
-    prefix: string,
-    exponent: number,
-}
-
 type Getter<T> = () => T;
-
-function getLargestLessThanOrEqualTo(list: number[], num: number): [number, number] {
-    // Example:
-    //  list = [1, 4, 7], num = 3 -> 1
-    //  list = [1, 4, 7], num = 4 -> 4
-    //  list = [1, 4, 7], num = 5 -> 4
-    //  list = [1, 4, 7], num = 100 -> 7
-    //  list = [1, 4, 7], num = -8 -> 1  Does not fail if less than lowest!
-    // Bisect a sorted, increasing array
-    // https://stackoverflow.com/a/12071013/20921535
-    let i = 0;
-    let end = list.length;
-    while ((end - i) > 1) {
-        const cursor = Math.floor((i + end) / 2);;
-        if (list[cursor] < num) {
-            i = cursor;
-        } else if (list[cursor] > num){
-            end = cursor;
-        } else {
-            i = end = cursor;
-        }
-    }
-    return [list[i], i];
-}
-
 type NonEmptyArray<T> = [T, ...T[]];
 
 class ScaleBarBase {
